requirements.txt
asgiref==3.7.2
Django==4.2.5
django-autoslug==1.9.9
django-jazzmin==2.6.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
